migrates webpack config to ts
